requirements-dev.txt
ruff>=0.9.4,<0.10.0
bandit>=1.8.2,<2.0.0
python-dotenv>=1.0.1,<2.0.0
